requirements.txt
gmpy
pycryptodome
bitstring
